Refactor out BaseEffect, Computation is common base now
